Per suggestions by ducky - now can invoke non scalatest test the same as the scalatest way   - import chisel3.tester.ChiselTester.test   - then just use `test(dutgen) { c => }`   - one example test changed to illustrate this way - Got rid of ChiselTesterHelper   - Adding functions from it as default implementations into BackendInterface - when using simple tester it creates a new test directory for every test   - s"chisel_test_${System.currentTimeMillis()}_$getCount"
