+++ src/main/scala/chisel3/tester/ChiselTester.scala
@@ -14,9 +14,8 @@ import scala.util.DynamicVariable
   * Used to run simple tests that do not require a scalatest environment in order to run
   * @param testName This will be used to generate a working directory in ./test_run_dir
   */
-class ChiselTester private (testName: String) extends Assertions with ChiselTesterHelper with TestEnvInterface {
+private class ChiselTester(testName: String) extends Assertions with TestEnvInterface {
   // Provide test fixture data as part of 'global' context during test runs
-
   val topFileName = Some(testName)
 
   private def runTest[T <: MultiIOModule](tester: BackendInstance[T])(testFn: T => Unit) {
@@ -46,19 +45,27 @@ class ChiselTester private (testName: String) extends Assertions with ChiselTest
   * @todo When Phases, Stages is implemented add ability to change testing options.
   */
 object ChiselTester {
+
+  private var count = 0
+
+  def getCount: Int = synchronized {
+    val current = count
+    count += 1
+    current
+  }
   /**
     * Run one test
     * @note every test should use a different name, it, suitably sanitized, is used as the subdirectory in the
     *       test_run_dir directory
-    * @param testName  The test name and where it's working directory
     * @param dutGen    The generator of the device under tests
     * @param testFn    The block of code that implements the test
     * @tparam T        The type of device, derived from dutGen
     */
-  def apply[T <: MultiIOModule]
-          (testName: String)
+  def test[T <: MultiIOModule]
           (dutGen: => T)
           (testFn: T => Unit): Unit = {
+
+    val testName = s"chisel_test_${System.currentTimeMillis()}_$getCount"
 
     val tester = new ChiselTester(testName)
     tester.test(dutGen)(testFn)
